mark session as stopped when terminated
This change marks the session as being stopped when a
'SessionTerminated' event is emitted from the Bloomberg Open API.  This
makes the blpapi-node API eaiser to use in the situation where the
connection goes down, the 'SessionTerminated' event is emitted and the
user of the API can immediately invoke '.destroy()'.  Otherwise, the
user of the API would need to  invoke '.stop()', but would never get
another 'SessionTerminated' event to then call '.destroy()'.