Support arbitrarily nested request objects.
This change enhances the generation of the `blpapi::Request` object to
support nested objects and arrays with arbitrary depths.  The algorithm
maintains a 1-to-1 mapping of the JavaScript request object to the C++
`blpapi::Request`; because of this,  the special handling for overrides
has changed and is not backwards compatible.

Overrides were originally specified as:
```json
{
  "overrides": {
    "<mnemonic or fieldId>": "<value>",
    "<mnemonic or fieldId>": "<value>"
  }
}
```

Now they are specified as:
```json
{
  "overrides": [{
    "fieldId": "<mneomnic or fieldId>",
    "value": "<value>"
  }, {
    "fieldId": "<mneomnic or fieldId>",
    "value": "<value>"
  }]
}
```